requirements.txt
speechrecognition>=3.10.0
pyttsx3>=2.90
gtts>=2.3.2
pygame>=2.1.2
requests>=2.28.0
openai>=1.0.0
python-dotenv>=0.21.0
Pillow>=9.0.0
